+++ frontend/src/components/settings/settingsComponents.js
@@ -30,101 +30,6 @@ import { Skeleton, SkeletonText, SkeletonCircle } from '../../lib/ui/Skeleton';
 import {useCategory} from "../../hooks/useCategory";
 import {AvatarRenderer} from "../../lib/AvatarRenderer";
 
-const getProjectIcon = (iconID) => {
-    switch (iconID) {
-        case 1: return <FiFolder size={24} />;
-        case 2: return <FiCodesandbox size={24} />;
-        case 3: return <FiCode size={24} />;
-        case 4: return <FiFileText size={24} />;
-        case 5: return <FiStar size={24} />;
-        case 6: return <FiBookmark size={24} />;
-        default: return <FiFolder size={24} />;
-    }
-};
-
-export const ErrorState = ({ error }) => (
-    <div className="p-6 flex flex-col items-center justify-center h-full">
-        <Card className="w-full max-w-md text-center">
-            <CardHeader>
-                <CardTitle className="text-2xl">
-                    Error <span className="text-oracleRed">Loading Settings</span>
-                </CardTitle>
-            </CardHeader>
-            <CardContent>
-                <div className="flex justify-center mb-6">
-                    <div className="w-16 h-16 rounded-full bg-red-50 flex items-center justify-center text-red-500">
-                        <FiAlertTriangle size={32} />
-                    </div>
-                </div>
-                <div className="bg-red-50 p-4 rounded-lg mb-6">
-                    <p className="text-red-600 font-medium">{error}</p>
-                </div>
-                <Button
-                    variant="outline"
-                    className="flex items-center gap-2"
-                    onClick={() => window.location.reload()}
-                >
-                    <FiRefreshCw size={16} />
-                    Retry
-                </Button>
-            </CardContent>
-        </Card>
-    </div>
-);
-
-export const NoProjectState = ({ title, message }) => (
-    <div className="p-6 flex flex-col items-center justify-center h-full">
-        <Card className="w-full max-w-md text-center">
-            <CardHeader>
-                <CardTitle className="text-2xl">
-                    No Project <span className="text-oracleRed">{title}</span>
-                </CardTitle>
-            </CardHeader>
-            <CardContent>
-                <div className="flex justify-center mb-6">
-                    <div className="w-16 h-16 rounded-full bg-gray-100 flex items-center justify-center text-gray-400">
-                        <FiFolder size={32} />
-                    </div>
-                </div>
-                <p className="text-gray-600 mb-6">
-                    {message}
-                </p>
-            </CardContent>
-        </Card>
-    </div>
-);
-
-export const ProjectHeader = ({ selectedProject, loading }) => (
-    <Card className="mb-6">
-        <CardHeader>
-            <div className={`flex items-center ${loading ? 'animate-pulse' : ''}`}>
-                <CardTitle>
-                    {loading ? (
-                        <div className="flex items-center">
-                            <SkeletonCircle size="md" />
-                            <div className="ml-3 w-48">
-                                <SkeletonText lines={1} />
-                            </div>
-                        </div>
-                    ) : (
-                        <div className="flex items-center">
-                            <div
-                                className="w-12 h-12 rounded-xl grid place-items-center text-white"
-                                style={{ backgroundColor: selectedProject?.color?.hexColor || '#808080' }}
-                            >
-                                {getProjectIcon(selectedProject?.icon)}
-                            </div>
-                            <h1 className="text-2xl font-bold px-3">
-                                {selectedProject?.projectName} <span className="text-oracleRed">Settings</span>
-                            </h1>
-                        </div>
-                    )}
-                </CardTitle>
-            </div>
-        </CardHeader>
-    </Card>
-);
-
 export const SettingsForm = ({
                                  loading,
                                  saving,
