+++ frontend/src/lib/ui/NoProject.jsx
@@ -1,0 +1,25 @@
+import {Card, CardContent, CardHeader, CardTitle} from "./Card";
+import {FiFolder} from "react-icons/fi";
+import React from "react";
+
+export const NoProjectState = ( { title, message } ) => (
+    <div className="p-6 flex flex-col items-center justify-center h-full">
+        <Card className="w-full max-w-md text-center">
+            <CardHeader>
+                <CardTitle className="text-2xl">
+                    No Project <span className="text-oracleRed">{title}</span>
+                </CardTitle>
+            </CardHeader>
+            <CardContent>
+                <div className="flex justify-center mb-6">
+                    <div className="w-16 h-16 rounded-full bg-gray-100 flex items-center justify-center text-gray-400">
+                        <FiFolder size={32} />
+                    </div>
+                </div>
+                <p className="text-gray-600 mb-6">
+                    { message }
+                </p>
+            </CardContent>
+        </Card>
+    </div>
+);
